create the "HashtagPage" folder
